Automatically switch the website to TLS
We definitely don’t want the user to accidentally interact with the
website insecurely.

Indeed, the website content could have been modified on-the-fly
if it is unprotected HTTP.

Of course, we cannot ensure that the initial HTTP request did not remove
the code to switch to the HTTPS version. That is what HSTS is for,
but sadly we cannot enable it on our own.